src/pyproject.toml
[tool.poetry]
name = "rag-chat"
version = "0.1.0"
description = ""
authors = ["adamjanasek <adamjanasek@o2.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
bandit = "1.7.8"
colorlog = "^6.8.2"
flake8 = "7.0.0"
isort = "^5.13.2"
langchain-openai = "^0.1.14"
mypy = "1.9.0"
pdfminer-six = "20231228"
pip-audit = "2.7.2"
pytest = "8.1.1"
pytest-cov = "5.0.0"
python = "^3.12"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
qdrant-client = "^1.10.0"
sentry-sdk = "^1.44.1"
xenon = "0.9.1"
openpyxl = "^3.1.2"
sanic = "^24.6.0"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
langchain-experimental = "^0.0.62"
pytest-asyncio = "^0.23.7"
pandas = "^2.2.2"
sanic-cors = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
